fix(review): address #43 review comments
Five issues flagged by reviewers (gemini-code-assist + chatgpt-codex):

1. **read_blob shouldn't enforce MAX_UPLOAD_BYTES** (HIGH, gemini).
   The cap is an ingress limit; enforcing it on read prevents loading
   blobs created under a higher historical limit or by other clients,
   and by the time bytes are in memory the check provides no
   memory-exhaustion protection anyway. Removed the read-side check
   with a comment explaining why.

2. **`now_ms` duplicated across handle modules** (MEDIUM, gemini).
   Promoted to `pub(crate) fn now_ms()` in handle/mod.rs; pages.rs and
   blobs.rs now import `super::now_ms`. Single source of truth.

3. **`create_space` passed empty strings as `Some("")`** (MEDIUM,
   gemini). The space manager expects `None` to trigger
   auto-generation/defaults; `Some("")` would persist an empty
   identifier. Use `(!s.is_empty()).then_some(s)` to map empty → None.

4. **`size: u32` risks silent truncation at 4GB** (MEDIUM, gemini).
   Widened `UploadBlobResultJs::size` and `ReadBlobResultJs::size` to
   `i64` (napi-rs emits as JS `number`, safe up to 2^53). The call
   sites cast `res.size as i64` instead of `as u32`. Future-proof
   without changing the JS-side type signature meaningfully.

5. **`upload_blob` didn't emit `DocumentBlobAdded` for doc uploads**
   (P2, codex). The gRPC path publishes this event right after
   `record_upload`; the new in-process path did not, so subscribers
   and downstream peer/cache workflows missed blobs uploaded via the
   addon for Yoopta/doc content. Mirrored the gRPC publish call
   inside `upload_blob` when `doc_id` is non-empty, with an info-level
   trace tagged `(in-process)` for log distinguishability.

Validation: cargo build -p soma-daemon -p soma-agentd -p soma-node;
cargo test on the same; pnpm --filter @soma/node run build:debug
regenerates index.d.ts with `size: number` on both blob types.

+++ backend/crates/soma-node/src/lib.rs
@@ -136,7 +136,7 @@ pub struct UploadBlobInputJs {
 #[napi(object)]
 pub struct UploadBlobResultJs {
     pub cid: String,
-    pub size: u32,
+    pub size: i64,
     pub mime: String,
     pub name: String,
 }
@@ -144,7 +144,7 @@ pub struct UploadBlobResultJs {
 #[napi(object)]
 pub struct ReadBlobResultJs {
     pub data: Buffer,
-    pub size: u32,
+    pub size: i64,
     pub mime: String,
 }
 
@@ -416,7 +416,7 @@ impl SomaHandle {
             .map_err(to_napi)?;
         Ok(UploadBlobResultJs {
             cid: res.cid,
-            size: res.size as u32,
+            size: res.size as i64,
             mime: res.mime,
             name: res.name,
         })
@@ -432,7 +432,7 @@ impl SomaHandle {
         let res = handle.read_blob(&space_id, &cid).await.map_err(to_napi)?;
         Ok(res.map(|r| ReadBlobResultJs {
             data: Buffer::from(r.data),
-            size: r.size as u32,
+            size: r.size as i64,
             mime: r.mime,
         }))
     }
